requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2
passlib[bcrypt]
python-dotenv
pyjwt
python-multipart
pyTelegramBotAPI
requests
python-dotenv